TestCases are only added once now.
TestCases were previously added once for each TestCase listed. This
meant you would have X * X TestCases in the Fixture where X is the
number of TestCases in the Fixture. Also of note is that all test cases
would be added for the current method, meaning some test cases for other
methods would be added to the current one.