AreEqual Double and Extended overloads did not check whether a value is NAN
If one of the values (actual or expected) passed to the method is Nan, then the framework did not check this situation, delegating SomeValue to provide a result.
As consequence the result may be random, since SameValue (neither AreEqual) checks the presence of a Nan value.
In my case I got that -Nan is equal to expected 100.