Fix FormatSetting not getting setup for versions lower than XE
While TFormatSettings cannot be created in non-XE version (it is a record) it is still passed in the .Format function so it needs to be setup.
The behavior that made me notice the issue is that the time taken for tests in the XML did not have decimals (they seemed to be in ms but in fact simply did not have a decimal separator).